Move Keybinding type definition to Keybinding.go
